Rename plugin folder & add dependency error handling
