+++ .gitignore
@@ -7,4 +7,4 @@ target/
 .classpath
 .project
 .settings/
-bin/
+APMServices/bin/
