+++ installAppScript
@@ -2,10 +2,10 @@
 
 echo -e "Installing application..."
 
-cf p apmServices -p ~/Repos/gradProject/APMServices/target/APMServices-1.0-SNAPSHOT.war
+cf p apmServices -p ~/Repos/gradProject/APMServices/target/APMServices-1.0-SNAPSHOT.war --no-start
 
 cf set-env apmServices staticAppUrl 'https://staticapp-hirable-phrasing.cfapps.io'
 
 cf set-env apmServices targetUrl 'https://api.run.pivotal.io'
 
-cf restage apmServices
+cf start apmServices
